+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Replace FollowAttacher.tagOnStart by tagOnAwake 
 - Increase Unity minimum version to 6000.1
 - Upgrade Unity Cinemachine package to 3.1.5